fix: move coupons position
