+++ src/components/customers/overview/CustomerCoupons.tsx
@@ -21,7 +21,7 @@ import { SectionHeader } from '~/styles/customer'
 import {
   AddCouponToCustomerDialog,
   AddCouponToCustomerDialogRef,
-} from './AddCouponToCustomerDialog'
+} from '../AddCouponToCustomerDialog'
 
 gql`
   fragment CustomerCoupon on AppliedCoupon {
@@ -85,7 +85,7 @@ export const CustomerCoupons = memo(() => {
     <>
       {!!(coupons || [])?.length && (
         <Container data-test="customer-coupon-container">
-          <SectionHeader variant="subhead">
+          <SectionHeader variant="subhead" $hideBottomShadow>
             {translate('text_628b8c693e464200e00e469d')}
             <Button
               variant="quaternary"
